new tests

+++ tests/test1.c
@@ -15,18 +15,18 @@ int main() {
   char m[100];
   readln(n);
   readln(m);
-  unsigned long long i = 1;
-  unsigned long long sum = 0;  // utilizamos o tipo de dados unsigned long long
-                               // para evitar overflow em sum
-  unsigned long long product = 1;  // msm caso de sum
+  int i = 1;
+  int sum = 0;      // utilizamos o tipo de dados int
+                    // para evitar overflow em sum
+  int product = 1;  // msm caso de sum
   int num = atoi(n);
   while (i <= num) {
     sum = sum + i;
     product = product * i;
     i = i + 1;
-    printf("soma: %llu \n", sum);  // utilizamos o especificador de formato %llu
-                                   // para imprimir valores unsigned long long
-    printf("produto: %llu \n", product);
+    printf("soma: %d \n", sum);  // utilizamos o especificador de formato %d
+                                 // para imprimir valores int
+    printf("produto: %d \n", product);
   }
   return 0;
 }
